// web/src/api/generated/model/trashItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * stocat
 * OpenAPI spec version: 0.0.1
 */
import type { TrashItemEncryptionMode } from "./trashItemEncryptionMode.ts";
import type { TrashItemKind } from "./trashItemKind.ts";

export interface TrashItem {
  created_at: string;
  encrypted_name?: string;
  encryption_mode: TrashItemEncryptionMode;
  expires_at: string;
  id: string;
  kind: TrashItemKind;
  library_id: string;
  library_name: string;
  name?: string;
  name_token?: string;
  parent_id: string;
  revision: number;
  trashed_at: string;
  updated_at: string;
}